refactor(db): change domain model IDs from Int to Long for overflow safety
- All ID fields changed to Long (64-bit) to prevent overflow
- Changed audioCurrentPos from Float to Double for precision
- Prevents integer overflow at >2B records

+++ backend/src/main/kotlin/com/lute/domain/Book.kt
@@ -4,13 +4,13 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Book(
-    val id: Int = 0,
-    val languageId: Int,
+    val id: Long = 0,
+    val languageId: Long,
     val title: String,
     val sourceURI: String? = null,
     val archived: Boolean = false,
-    val currentTextId: Int = 0,
+    val currentTextId: Long = 0,
     val audioFilename: String? = null,
-    val audioCurrentPos: Float? = null,
+    val audioCurrentPos: Double? = null,
     val audioBookmarks: String? = null,
 )
